requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
tiktoken
langchain-core
PyYAML
langchain-community
